requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Handling
aiofiles==23.2.0
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Data Validation
pydantic==2.5.0
email-validator==2.1.0

# Environment
python-dotenv==1.0.0

# Audio Processing (for speaking tests)
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1

# AI/ML Libraries
openai==1.3.0
anthropic==0.8.0
google-generativeai==0.3.2

# Background Tasks
celery==5.3.4
redis==5.0.1

# Image Processing (for listening maps/diagrams)
Pillow==10.1.0

# PDF Processing (if needed for reading passages)
PyPDF2==3.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0